+++ Software/Core/OptimalControl/include/OptimalControl/OptimalControlProblem.hpp
@@ -83,16 +83,16 @@ namespace OptimalControl
         Eigen::MatrixXd Q_; // State Weights
         Eigen::MatrixXd R_; // Input Weights
 
-        int num_states_; // Number of System States
-        int num_inputs_; // Number of System Inputs
+        int64_t num_states_; // Number of System States
+        int64_t num_inputs_; // Number of System Inputs
 
-        int N_; // Number of Prediction Steps
+        int64_t N_; // Number of Prediction Steps
 
         double T_s_; // Sample Time
         double T_;   // Horizon Length
 
-        int max_iterations_;    // Max Iterations for Solver
-        int solver_iterations_; // Total number of Solver iterations for solution
+        int64_t max_iterations_;    // Max Iterations for Solver
+        int64_t solver_iterations_; // Total number of Solver iterations for solution
         double solver_time_;    // Total time for Solver to compute a solution
 
         bool solved_;
